wip: Show only bootable ISOs in dialog
This doesn't really work because shared-mime-info doesn't know about our
'x-content/bootable-media' mime-type.

I talked to Bastien about making xdg-mime aware of this mime-type but I
was told that xdg-mime is supposed to be able to guess mime-type by only
sniffing the first few kilobytes, where as we need to skip at least 34K
to detect if El torito header is present on ISO (which is what implies a
bootable ISO).