util-app: Add get_system_virt_connection()
This moves creation of system libvirt connection to utils. We'll be
starting to re-use the same connection for fetching the IP address of
guests in a following patch.

This also implies that this connection is never properly unrefed/freed.
Keeping in mind that this instance was kept around for lifetime of Boxes
already and that libvirt is fully capable of cleaning up after apps, this
is not such a bad thing.

https://bugzilla.gnome.org/show_bug.cgi?id=744004