machine: Add virtual update_info()
In commit b1aaa70, we consolidated the setting of 'info' to some extent
by ensuring the this property is only set by
LibvirtMachine.update_info(). This patch takes that effort further by
making LibvirtMachine.update_info() a virtual method of the base Machine
class.

This also implies that Machine no longer keep the existing status string
and instead of concatenating new strings, it simply overrides it. This
is the right thing to do IMO since no critical information is kept in
this property and so far we haven't had the need to keep two different
pieces of information at the same time in it.

https://bugzilla.gnome.org/show_bug.cgi?id=744004