wizard: Ignore unimportable VMs early
The problem is that libvirt VMs with an invalid disk image
have the option to be imported in the wizard.

To fix this, check the existence of the VMs' disk images and
ensure their readability early on. Simply ignore the VMs with
an invalid disk image.

Remove the old readability check.

https://bugzilla.gnome.org/show_bug.cgi?id=754500