libvirt-machine-props: Fetch snapshots at construction
Instead of fetching snapshots at runtime, do so as part of
LibvirtMachineProperties construction (which in turn happens as part of
LibvirtMachine construction). This implies we re-fetch them each time a
snapshot is added or removed to keep the list up2date.

This change is needed because fetching of snapshots is an async
operation (and we want to keep it that way to avoid hanging the UI) and
we'll need access to snapshots during the sync operation of properties
construction, in a following patch.