Add PropertiesSnapshots widget
This is almost identical to SnapshotsProperty but is a widget itself and
extends PropertiesPageWidget. This will enable us to change
IPropertiesProvider.get_properties() to return a PropertiesPageWidget
rather than a list of Property objects in a following patch and
ultimately to drop Property hierarchy in favour of page-specific widget
implementation.