meson: Fix post install script
meson uses a post install script for compiling gsettings schemas,
updating the icon cache and also updating the desktop database.
However, it assumes that the data directory is the share directory
under prefix, which might not be correct because the user can choose
a different directory.

This patch uses the data directory used by meson by passing it to
the post install script.

https://bugzilla.gnome.org/show_bug.cgi?id=791421